fix: Don't unqualify fields
Unqualified fields were used when we did not embed the search in the inner root query and needed to remove table names as the conditions were being evaluated on the outer query.
With the inner query, we shouldn't unqualify, as it can lead to ambiguous targeting.

This change removes the de-qualifying that was previously required.